refactor: Relocate port interfaces from `src/ports` to `src/domain/ports` and update all module references.
