+++ src/adapters/workspace_locator/current_directory_locator.rs
@@ -1,5 +1,5 @@
 use crate::domain::error::AppError;
-use crate::ports::WorkspaceLocator;
+use crate::domain::ports::WorkspaceLocator;
 use std::path::PathBuf;
 
 #[derive(Debug, Default, Clone)]
