+++ src/domain/mod.rs
@@ -1,5 +1,6 @@
 pub mod context_file;
 pub mod error;
+pub mod ports;
 pub mod snippet;
 
 pub use error::AppError;
