+++ src/domain/ports/clipboard.rs
