+++ src/domain/ports/context_file_store.rs
@@ -1,5 +1,6 @@
-use crate::domain::error::AppError;
 use std::path::{Path, PathBuf};
+
+use crate::domain::error::AppError;
 
 #[derive(Debug, Clone)]
 pub struct ContextWriteStatus {
