+++ src/domain/ports/snippet_catalog.rs
