+++ src/domain/ports/workspace_locator.rs
@@ -1,5 +1,6 @@
-use crate::domain::error::AppError;
 use std::path::PathBuf;
+
+use crate::domain::error::AppError;
 
 pub trait WorkspaceLocator {
     fn find_workspace_root(&self) -> Result<PathBuf, AppError>;
