+++ src/testing/ports/in_memory_catalog.rs
@@ -1,6 +1,6 @@
 use crate::domain::error::AppError;
+use crate::domain::ports::SnippetCatalog;
 use crate::domain::snippet::SnippetEntry;
-use crate::ports::SnippetCatalog;
 
 pub struct InMemoryCatalog {
     entries: Vec<SnippetEntry>,
