+++ src/testing/ports/in_memory_context_store.rs
@@ -1,5 +1,5 @@
 use crate::domain::error::AppError;
-use crate::ports::{ContextFileStore, ContextWriteStatus};
+use crate::domain::ports::{ContextFileStore, ContextWriteStatus};
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
